python-analyzer/requirements.txt
flask==2.0.1
flask-cors==3.0.10
nltk==3.6.2
scikit-learn==0.24.2
pandas==1.3.3
numpy==1.21.2
konlpy==0.5.2
kafka-python==2.0.2
python-dotenv==0.19.0 